requirements.txt
aiogram==3.*
psycopg2-binary
openai
python-dateutil